Move magic traits queries to rustc::traits::drop.
